added login refactor
