+++ web/src/pages/login/ConsentBanner.js
